// src/api/generated/model/rewardsControllerGetRewardsParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

export type RewardsControllerGetRewardsParams = {
/**
 * Номер страницы (от 1)
 */
page?: number;
/**
 * Размер страницы
 */
size?: number;
/**
 * Включать удалённые награды
 */
includeDeleted?: boolean;
};